update axes cycler when Call style attribute is changed
only when Call has been assigned to an axes.  If it hasn't, then the axes will pull the style attribute when it is added and append that to the cycler.

The tmp values in the cycler can add extra entries, but they will clear on the next draw and be in sync again.