obey highlight_linestyle for plot itself
before highlight_linestyle was only applied to axvline and axhline when i was referring to an existing dimension.  Now, when i is not an existing dimension, but i is instead used to SELECT from a set, the highlight_linestyle will be obeyed.  In this way, you can set linestyle='none', highlight_linestyle='solid' to show only the selected objects.