ignore pad_aspect for animations
temporary solution to #18.  Fixing this once and for all will require writing a custom callback (in callbacks.py) within the resize/zoom to manually pad the datalimits instead of using matplotlib's native behavior for set_aspect like we do now.  But with this temporary fix, pad_aspect=True will be ignored for animations, but the automatic axes limits will still be respected and set correctly.